% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all_stations_all_dates.R
\name{all_stations_all_dates}
\alias{all_stations_all_dates}
\title{All Stations, All Dates}
\usage{
all_stations_all_dates(town, pollutant)
}
\arguments{
\item{town}{Biomass Study area}

\item{pollutant}{you got it}
}
\value{
text for a query
}
\description{
All Stations, All Dates
}

