% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BiosmokeValidatedEvents-package.r
\docType{package}
\name{BiosmokeValidatedEvents}
\alias{BiosmokeValidatedEvents}
\alias{BiosmokeValidatedEvents-package}
\title{BiosmokeValidatedEvents}
\description{
BiosmokeValidatedEvents
}

