% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sites_todo.R
\name{citywide_av}
\alias{citywide_av}
\title{city wide average}
\usage{
citywide_av(town, poll, stat)
}
\arguments{
\item{town}{}

\item{poll}{}

\item{stat}{}
}
\value{
nothing to R, this creates things in the database
}
\description{
city wide average
}

