% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{impute for each site}
\usage{
impute(sitelist = c("SouthLake", "Duncraig"), town = "PERTH",
  poll = "pm10", stat = "av", maxdate = "2007-12-31")
}
\arguments{
\item{sitelist}{sites}

\item{town}{town}

\item{poll}{pollutant}

\item{stat}{statistical unit as per avg or max}

\item{maxdate}{the end of the time series}
}
\value{
database table
}
\description{
impute for each site
}

