% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mindates_pollutant.R
\name{mindates_pollutant}
\alias{mindates_pollutant}
\title{Minimum Date a Pollutant is observed from}
\usage{
mindates_pollutant(town = "perth", pollutant = "pm10_av")
}
\arguments{
\item{town}{the Biomass Study Town in question}

\item{pollutant}{you got it}
}
\value{
text for a SQL query
}
\description{
Minimum Date a Pollutant is observed from
}

