% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sites_todo.R
\name{n_missing}
\alias{n_missing}
\title{number missing}
\usage{
n_missing(town, poll, thresh = 0.05)
}
\arguments{
\item{town}{the one to do}

\item{poll}{pollutant}

\item{thresh}{theshold below which we will do it}
}
\value{
nmissing is a message like 'go for it'
}
\description{
number missing
}

