% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sites_todo.R
\name{sites_todo}
\alias{sites_todo}
\title{sites with potential}
\usage{
sites_todo(town, mindate, maxdate = "'2007-12-31'", threshold = 0.7, poll,
  stat)
}
\arguments{
\item{town}{}

\item{mindate}{}

\item{maxdate}{}

\item{threshold}{}

\item{poll}{}

\item{stat}{}
}
\value{
text for a sql query
}
\description{
sites with potential
}

