% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sites_todo.R
\name{stitch_together}
\alias{stitch_together}
\title{put all the bits together}
\usage{
stitch_together(poll = polls[5, 3])
}
\arguments{
\item{poll}{pollutant}
}
\value{
tables in the database
}
\description{
put all the bits together
}

