% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stitch_together.R
\name{stitch_together}
\alias{stitch_together}
\title{put all the bits together}
\usage{
stitch_together(poll = polls[5, 3], stat = "av")
}
\arguments{
\item{poll}{pollutant}

\item{stat}{av or max}
}
\value{
tables in the database
}
\description{
put all the bits together
}

